chore: move out test setup function
